requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
boilerpy3==1.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
elasticsearch==7.9.0
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
haystack-ai==2.0.0b3
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
openai==0.28.1
opensearch-haystack==0.1.1
opensearch-py==2.4.2
pandas==2.1.4
posthog==3.1.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.32.0.post1
tenacity==8.2.3
tqdm==4.66.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
uvicorn==0.24.0
